src/streamlit/confluence_agent.py
"""
Confluence Agent - Documentation and knowledge base
Tools for searching and retrieving Confluence documentation.
"""

import os

from strands import Agent, tool
from strands.models import BedrockModel

# Configuration
CONFLUENCE_URL = os.environ.get("CONFLUENCE_URL", "https://mrrobot.atlassian.net")
CONFLUENCE_TOKEN = os.environ.get("CONFLUENCE_TOKEN", "")

# ============================================================================
# TOOLS
# ============================================================================


@tool
def search_confluence(query: str, space_key: str = "", limit: int = 20) -> str:
    """Search Confluence for pages matching a query.

    Args:
        query: Search query string
        space_key: Limit to specific space (e.g., 'DEV', 'OPS', 'PROD')
        limit: Maximum number of results
    """
    # TODO: Implement Confluence API call
    return "TODO: Implement search_confluence"


@tool
def get_page_content(page_id: str) -> str:
    """Get the content of a specific Confluence page.

    Args:
        page_id: Confluence page ID
    """
    # TODO: Implement Confluence API call
    return "TODO: Implement get_page_content"


@tool
def get_page_by_title(title: str, space_key: str) -> str:
    """Get a Confluence page by its title.

    Args:
        title: Page title
        space_key: Space key where the page lives
    """
    # TODO: Implement Confluence API call
    return "TODO: Implement get_page_by_title"


@tool
def list_spaces() -> str:
    """List all available Confluence spaces."""
    # TODO: Implement Confluence API call
    return "TODO: Implement list_spaces"


@tool
def get_space_pages(space_key: str, limit: int = 50) -> str:
    """List pages in a specific Confluence space.

    Args:
        space_key: Space key (e.g., 'DEV', 'OPS')
        limit: Maximum number of pages to return
    """
    # TODO: Implement Confluence API call
    return "TODO: Implement get_space_pages"


@tool
def get_recent_updates(space_key: str = "", days_back: int = 7, limit: int = 20) -> str:
    """Get recently updated pages.

    Args:
        space_key: Limit to specific space (optional)
        days_back: How many days back to look
        limit: Maximum number of results
    """
    # TODO: Implement Confluence API call
    return "TODO: Implement get_recent_updates"


@tool
def search_by_label(label: str, space_key: str = "", limit: int = 20) -> str:
    """Find pages with a specific label.

    Args:
        label: Label to search for
        space_key: Limit to specific space (optional)
        limit: Maximum number of results
    """
    # TODO: Implement Confluence API call
    return "TODO: Implement search_by_label"


# Export tools list
CONFLUENCE_TOOLS = [
    search_confluence,
    get_page_content,
    get_page_by_title,
    list_spaces,
    get_space_pages,
    get_recent_updates,
    search_by_label,
]

# System prompt
SYSTEM_PROMPT = """You are a Confluence Documentation Assistant for MrRobot.

You help the team find and retrieve documentation:
- Search for relevant documentation
- Retrieve page content
- Find recently updated docs
- Navigate spaces and labels

AVAILABLE TOOLS:
1. search_confluence - Full-text search across Confluence
2. get_page_content - Get content of a specific page
3. get_page_by_title - Find page by exact title
4. list_spaces - List all spaces
5. get_space_pages - List pages in a space
6. get_recent_updates - Find recently modified pages
7. search_by_label - Find pages by label

COMMON SPACES:
- DEV: Development documentation
- OPS: Operations runbooks
- ARCH: Architecture decisions
- ONBOARD: Onboarding guides

When searching, try multiple approaches if the first doesn't find results.
"""


# Create agent
def create_confluence_agent():
    model = BedrockModel(model_id="us.anthropic.claude-sonnet-4-20250514-v1:0", region_name="us-west-2")
    return Agent(model=model, tools=CONFLUENCE_TOOLS, system_prompt=SYSTEM_PROMPT)


confluence_agent = None  # Lazy initialization
