requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
plotly>=5.0.0
streamlit>=1.28.0

# Strands Agent SDK
strands-agents>=1.0.0

# AWS SDK
boto3>=1.26.0
botocore>=1.29.0

# HTTP and Flask
requests>=2.28.0
flask>=2.3.0

# MCP Server SDK (official Anthropic SDK)
mcp>=1.0.0

# ASGI server for MCP HTTP transport
uvicorn>=0.24.0
starlette>=0.27.0

# Slack Bot SDK
slack-bolt>=1.18.0
slack-sdk>=3.21.0
